test(v0.14): keep synthetic row count strongly typed

+++ ui/musicark_ui/test/v014_performance_hardening_test.dart
@@ -22,7 +22,7 @@ class _LargeLocalBridge extends FakeMusicArkBridge {
     List<int>? rootIds,
   }) async {
     requestedLimits.add(limit);
-    final available = (10_000 - offset).clamp(0, limit);
+    final available = (10_000 - offset).clamp(0, limit).toInt();
     return {
       'count': 10_000,
       'limit': limit,
